+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Misti API to execute the driver programmatically
 - `ArgCopyMutation`: Report once per function: Issue [#150](https://github.com/nowarp/misti/issues/150)
 - `ArgCopyMutation`: Don't report arguments returned from the function: Issue [#149](https://github.com/nowarp/misti/issues/149)
+- Rename CLI options: `--suppress` -> `--disable-detectors`; `--detectors` -> `--enable-detectors`
 
 ## [0.3.1] - 2024-09-24